First successful build
Finally got ibcore-root to build.  There are disabled tests, so it
is not ready for GA, but at least it can be worked with.

No guarantees are made that the tests actually allow us to see if these
modules can be used for their intended purposes.  Some integration
testing will be very necessary.