It all compiles
It's TOTALLY not correct.  All the modules are wrong.  All the imports are
probably at least somewhat wrong because all the modules are wrong.

It doesn't follow the prescribed-by-sane-people notions of module naming.
It doesn't follow any decently mannered person or their ways.

It just compiles and for that, I am ashamed.

But it compiles